requirements.txt
mcp
yfinance
pandas
numpy
scipy
pandas_ta
gnews
newsapi-python
pycoingecko
alpaca-py
python-dotenv
modal>=0.63.0
requests
pydantic>=2.12.3
gradio>=6.0.1
matplotlib>=3.10.7
seaborn>=0.13.2
mplfinance>=0.12.10b0
pillow>=12.0.0
